change the random number double to integer and then it save as string

+++ Backend/src/main/java/com/raj/library/Service/OtpService.java
@@ -18,7 +18,7 @@ public class OtpService {
     public boolean generateOTP(String username){
         String generatedOTP = "";
         for(int i=0;i<=otpLength;i++){
-            generatedOTP += String.valueOf((Math.random()*9)+1);
+            generatedOTP += String.valueOf((int)(Math.random()*9)+1);
         }
         otpMapping.put(username,generatedOTP);
         ScheduledExecutorService scheduledExecutorService = Executors.newScheduledThreadPool(1);
